# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * World Capo API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Flag {
  id: number;
  teamId: string;
  x: number;
  y: number;
  createdAt: string;
}

export interface TeamFlagCount {
  teamId: string;
  count: number;
}

export interface PaymentIntentInput {
  teamId: string;
  x: number;
  y: number;
  /** @nullable */
  email?: string | null;
}

export interface PaymentIntentResult {
  clientSecret: string;
  paymentIntentId: string;
}

export interface WebhookAck {
  received: boolean;
}

export interface AdminStats {
  totalRevenue: number;
  totalFlags: number;
  totalPayments: number;
  successRate: number;
  revenueToday: number;
  flagsToday: number;
}

export interface AdminPayment {
  id: number;
  stripePaymentIntentId: string;
  amountCents: number;
  currency: string;
  status: string;
  teamId: string;
  createdAt: string;
}

export interface AdminPaymentList {
  payments: AdminPayment[];
  total: number;
  page: number;
  limit: number;
}

export interface TeamBreakdown {
  teamId: string;
  flagCount: number;
  revenue: number;
}

export interface ConsentInput {
  sessionId: string;
  analytics: boolean;
  marketing: boolean;
}

export interface ConsentResult {
  saved: boolean;
}

export interface ErasureInput {
  email: string;
}

export interface ErasureResult {
  acknowledged: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListAdminPaymentsParams = {
page?: number;
limit?: number;
status?: string;
};