# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: World Capo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/intent:
    post:
      operationId: createPaymentIntent
      tags: [payments]
      summary: Create a Stripe PaymentIntent for flag placement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentIntentInput"
      responses:
        "200":
          description: PaymentIntent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /gdpr/consent:
    post:
      operationId: recordConsent
      tags: [gdpr]
      summary: Record user cookie/analytics consent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsentInput"
      responses:
        "200":
          description: Consent recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsentResult"
  /gdpr/erasure:
    post:
      operationId: requestErasure
      tags: [gdpr]
      summary: GDPR right-to-erasure request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErasureInput"
      responses:
        "200":
          description: Erasure request acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErasureResult"
components:
  schemas:
...
    PaymentIntentInput:
      type: object
      required: [teamId, x, y]
      properties:
        teamId:
          type: string
        x:
          type: number
        y:
          type: number
        email:
          type: ["string", "null"]
    PaymentIntentResult:
      type: object
      required: [clientSecret, paymentIntentId]
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
...
    ConsentInput:
      type: object
      required: [sessionId, analytics, marketing]
      properties:
        sessionId:
          type: string
        analytics:
          type: boolean
        marketing:
          type: boolean
    ConsentResult:
      type: object
      required: [saved]
      properties:
        saved:
          type: boolean
    ErasureInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ErasureResult:
      type: object
      required: [acknowledged, message]
      properties:
        acknowledged:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string