# lib/api-zod/src/generated/types/consentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * World Capo API
 * OpenAPI spec version: 0.1.0
 */

export interface ConsentInput {
  sessionId: string;
  analytics: boolean;
  marketing: boolean;
}